feat(item-icon.vue): refactor item icon component and add item name tooltip
